Refactoring - moved common code to base client
1) The client hierarchy continues to evolve - moved common, lower level
code to base ReadmillClient class and made the base class abstract - we
don't want users to directly instantiate base client but use one of the
more derived versions. The interface for getting to the specifc client
objects will be further designed later.

2) Minor code clean up in tests - We don't need the Continuation for
tests since we don't mind blocking on the call to Task.Result here