Porting ReadmillClient to Windows Phone 7
1) Ported ReadmillClient to Windows Phone 7 (.csproj ReadmillClient7).
This required several design hacks / changed:
1.1) Tasks are not supported out of box on Windows Phone 7 so added a
nuget package for an open source library. This is a port of mono.net
version of the Task library for Windows Phone and is supposed to be
feature compatible with the official .Net version.
1.2) HttpClient is not available for Windows Phone 7 so switched to
using WebClient and HttpWebRequest in that version in
ReadmillClientBase. Wrapped the WebClient and HttpWebRequest calls with
Tasks based on the standard pattern so that the publicly exposed API
doesn't change.
1.3) UriTemplate class doesn't work on Windows Phone 7 so create a
local, bare-bones implementation for the same.
All the above changes need to be throughly tested.

2) Added scaffolding code for 'Read7' - a Windows Phone app that uses
the ReadmillClient.