Removed todo comment
The comment asked if the progress of a Ping needs to be more
than the progress of the last session. The answer is no =)

Progress for pings is just a snapshot value of where the user is at the
time of the ping.

The highest progress value of all pings sent for a given Period
(determined by 'sessionId') will be used as the progress of
that Period.